requirements.txt
psutil
python-ptrace